feat(test): Ajout des tests d'integration avec testContainers - restTemplate

+++ src/main/java/sn/fr/samagp/services/impl/AnnonceServiceImp.java
@@ -3,6 +3,7 @@ package sn.fr.samagp.services.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import sn.fr.samagp.mapper.AnnonceMapper;
+import sn.fr.samagp.mapper.ClientMapper;
 import sn.fr.samagp.repository.AnnonceRepository;
 import sn.fr.samagp.repository.dto.AnnonceDTO;
 import sn.fr.samagp.repository.model.Annonce;
@@ -10,6 +11,7 @@ import sn.fr.samagp.repository.response.AnnonceResponse;
 import sn.fr.samagp.services.inter.IAnnonce;
 import sn.fr.samagp.validator.AnnonceValidator;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -21,24 +23,26 @@ public class AnnonceServiceImp implements IAnnonce {
     private final AnnonceRepository annonceRepository;
     private final AnnonceMapper annonceMapper;
     private final AnnonceValidator annonceValidator;
+    private final ClientMapper clientMapper;
 
-    public AnnonceServiceImp(AnnonceRepository annonceRepository, AnnonceMapper annonceMapper, AnnonceValidator annonceValidator) {
+    public AnnonceServiceImp(AnnonceRepository annonceRepository, AnnonceMapper annonceMapper, AnnonceValidator annonceValidator, ClientMapper clientMapper) {
         this.annonceRepository = annonceRepository;
         this.annonceMapper = annonceMapper;
         this.annonceValidator = annonceValidator;
+        this.clientMapper = clientMapper;
     }
 
 
     @Override
-    public AnnonceDTO createAnnonce(AnnonceDTO annonceDTO) {
+    public AnnonceResponse createAnnonce(AnnonceDTO annonceDTO) {
         Annonce annonce = annonceMapper.toEntity(annonceDTO);
         annonceValidator.validate(annonce);
-        return annonceMapper.toDto(annonceRepository.save(annonce));
+        return annonceMapper.toResponse(annonceRepository.save(annonce));
     }
 
     @Override
     public Optional<AnnonceResponse> getAnnonceById(UUID id) {
-        return annonceRepository.findById(id).map(annonceMapper::toResponse);
+        return annonceRepository.findByAltId(id).map(annonceMapper::toResponse);
     }
 
     @Override
@@ -52,19 +56,26 @@ public class AnnonceServiceImp implements IAnnonce {
     public AnnonceDTO updateAnnonce(UUID id, AnnonceDTO annonceDTO) {
         return annonceRepository.findById(id)
                 .map(existingAnnonce -> {
-//                    existingAnnonce.setItineraire(itineraireMapper.toEntity(annonceDTO.itineraireDTO()));
-//                    existingAnnonce.setItineraireDetailsDepart(annonceDTO.itineraireDetailsDepart());
-//                    existingAnnonce.setItineraireDetailsArrive(annonceDTO.itineraireDetailsArrive());
-//                    existingAnnonce.setDateDepart(annonceDTO.dateDepart());
-//                    existingAnnonce.setDateArrive(annonceDTO.dateArrive());
-//                    existingAnnonce.setUpdatedAt(LocalDateTime.now());
+                    existingAnnonce.setItineraireDetailsDepart(annonceDTO.getItineraireDetailsDepart());
+                    existingAnnonce.setItineraireDetailsArrive(annonceDTO.getItineraireDetailsArrive());
+                    existingAnnonce.setDateDepart(annonceDTO.getDateDepart());
+                    existingAnnonce.setDateArrive(annonceDTO.getDateArrive());
+                    existingAnnonce.setUpdatedAt(LocalDateTime.now());
+                    existingAnnonce.setDescription(annonceDTO.getDescription());
+                    existingAnnonce.setClient(clientMapper.toEntity(annonceDTO.getClientDTO()));
                     return annonceMapper.toDto(annonceRepository.save(existingAnnonce));
                 }).orElseThrow(() -> new RuntimeException("Annonce non trouvée!"));
     }
 
     @Override
     public void deleteAnnonce(UUID id) {
-        annonceRepository.deleteById(id);
+        Optional<Annonce> annonceToDelete = annonceRepository.findById(id);
+        if (annonceToDelete.isPresent()) {
+            System.out.println("************ Suppression de l'annonce : " + annonceToDelete.get());
+            annonceRepository.deleteById(id);
+        } else {
+            System.out.println("************** Annonce avec l'ID " + id + " introuvable !");
+        }
     }
 
     @Override
